+++ packages/brick-utils/src/index.ts
@@ -20,3 +20,6 @@ export * from "./scanCustomApisInStoryboard";
 export * from "./smartDisplayForEvaluableString";
 export { scanStoryboard } from "./scanStoryboard";
 export * from "./JsonStorage";
+export * from "./builder";
+export * from "./deepFreeze";
+export * from "./trackContext";