+++ packages/sdk-scripts/CHANGELOG.md
@@ -3,13 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.6.10](https://github.com/easyops-cn/next-core/compare/@next-core/sdk-scripts@2.6.9...@next-core/sdk-scripts@2.6.10) (2021-07-02)
+## [2.6.11](https://github.com/easyops-cn/next-core/compare/@next-core/sdk-scripts@2.6.10...@next-core/sdk-scripts@2.6.11) (2021-07-23)
 
 **Note:** Version bump only for package @next-core/sdk-scripts
 
 
 
 
+
+## [2.6.10](https://github.com/easyops-cn/next-core/compare/@next-core/sdk-scripts@2.6.9...@next-core/sdk-scripts@2.6.10) (2021-07-02)
+
+**Note:** Version bump only for package @next-core/sdk-scripts
 
 ## [2.6.9](https://github.com/easyops-cn/next-core/compare/@next-core/sdk-scripts@2.6.8...@next-core/sdk-scripts@2.6.9) (2021-06-23)